refactor(utils): utils module refactor
